+++ backend/src/middlewares/authorization.middleware.js
@@ -1,0 +1,40 @@
+"use strict";
+// Autorizacion - Comprobar el rol del usuario
+const User = require("../models/user.model.js");
+const Role = require("../models/role.model.js");
+const { respondError } = require("../utils/resHandler.js");
+const { handleError } = require("../utils/errorHandler.js");
+
+/**
+ * @name isAdmin
+ * @description Comprueba si el usuario es administrador
+ * @param req {Request}
+ * @param res {Response}
+ * @param next {Function}
+ * @returns {Promise<void>}
+ * @throws {Error}
+ */
+async function isAdmin(req, res, next) {
+  try {
+    const user = await User.findOne({ email: req.email });
+    const roles = await Role.find({ _id: { $in: user.roles } });
+    for (let i = 0; i < roles.length; i++) {
+      if (roles[i].name === "admin") {
+        next();
+        return;
+      }
+    }
+    return respondError(
+      req,
+      res,
+      401,
+      "Se requiere un rol de administrador para realizar esta acción",
+    );
+  } catch (error) {
+    handleError(error, "authorization.middleware -> isAdmin");
+  }
+}
+
+module.exports = {
+  isAdmin,
+};
